+++ src/FamilyHubs.SharedKernel.Razor/FamilyHubsUi/Options/FhHealthCheckOptions.cs
@@ -1,28 +1,6 @@
-﻿
+﻿using FamilyHubs.SharedKernel.Razor.FamilyHubsUi.Options.HealthCheck;
+
 namespace FamilyHubs.SharedKernel.Razor.FamilyHubsUi.Options;
-
-public class HealthCheckDatabaseOptions
-{
-    public enum DatabaseType
-    {
-        SqlServer
-    }
-
-    /// <summary>
-    /// The ConnectionString.
-    /// </summary>
-    public string? ConnectionString { get; set; }
-
-    /// <summary>
-    /// If supplied, the ConnectionString is populated from the config value found at the given config key.
-    /// </summary>
-    /// <example>
-    /// "ConnectionStrings:SharedKernelConnection"
-    /// </example>
-    public string? ConfigConnectionString { get; set; }
-
-    public DatabaseType Type { get; set; } = DatabaseType.SqlServer;
-}
 
 public class HealthCheckUrlOptions
 {
