+++ src/FamilyHubs.SharedKernel.Razor/FamilyHubsUi/Options/HealthCheck/HealthCheckDatabaseOptions.cs
@@ -1,0 +1,25 @@
+﻿
+namespace FamilyHubs.SharedKernel.Razor.FamilyHubsUi.Options.HealthCheck;
+
+public class HealthCheckDatabaseOptions
+{
+    public enum DatabaseType
+    {
+        SqlServer
+    }
+
+    /// <summary>
+    /// The ConnectionString.
+    /// </summary>
+    public string? ConnectionString { get; set; }
+
+    /// <summary>
+    /// If supplied, the ConnectionString is populated from the config value found at the given config key.
+    /// </summary>
+    /// <example>
+    /// "ConnectionStrings:SharedKernelConnection"
+    /// </example>
+    public string? ConfigConnectionString { get; set; }
+
+    public DatabaseType Type { get; set; } = DatabaseType.SqlServer;
+}
